vector-db/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymilvus==2.3.4
pinecone-client==2.2.4
numpy==1.24.3
pydantic==2.4.2
python-dotenv==1.0.0
openai==1.3.0
python-multipart==0.0.6 